Add mode entry points to our preset (#4669)
