+++ packages/babel-preset-react-app/package.json
@@ -9,7 +9,11 @@
   },
   "files": [
     "index.js",
-    "dependencies.js"
+    "create.js",
+    "dependencies.js",
+    "dev.js",
+    "prod.js",
+    "test.js"
   ],
   "dependencies": {
     "@babel/core": "7.0.0-beta.46",
